requirements.txt
black==21.12b0
Django==3.2.0
django-cors-middleware==1.5.0
django-countries==7.2.1
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth[with_social]==0.9.5
djangorestframework
djangorestframework-filters
gevent==21.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pre-commit==2.12.1
pylint==2.7.2
pylint-django==2.4.4
python-decouple==3.4
